refactor: extract HookInstallationCoordinator from AppModel
Move hook/usage installation management into a dedicated @Observable
coordinator. This extracts ~430 lines from AppModel including:
- 3 installation managers (Codex hooks, Claude hooks, Claude status line)
- 5 status properties and 3 busy flags
- 2 usage monitoring tasks
- 14 methods (refresh/install/uninstall/update)
- All computed display properties (status titles, summaries, usage text)

AppModel retains forwarding computed properties and methods so the
View layer requires zero changes.
